+++ src/load/old/item-flag-types-savefile10.h
