+++ src/load/old/monster-flag-types-savefile10.h
