+++ src/load/old/monster-loader-savefile10.cpp
@@ -2,7 +2,7 @@
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
 #include "load/old/load-v1-5-0.h"
-#include "load/old/savedata10-monster-flag-types.h"
+#include "load/old/monster-flag-types-savefile10.h"
 #include "system/monster-type-definition.h"
 #include "system/player-type-definition.h"
 #include "util/bit-flags-calculator.h"
