+++ src/save/monster-writer.cpp
@@ -1,5 +1,5 @@
 ﻿#include "save/monster-writer.h"
-#include "load/old/savedata10-monster-flag-types.h"
+#include "load/old/monster-flag-types-savefile10.h"
 #include "monster-race/monster-race.h"
 #include "monster/monster-info.h"
 #include "monster/monster-status.h"
